Refactor: 게시물 작성 및 수정 컴포넌트 로직분리
- 두 개의 컴포넌트에서 겹치는 로직이 있어 util폴더에 parseImg 파일로 로직을 구분하였습니다.
- 두 개의 컴포넌트를 사용해서 페이지를 구성할 지, 하나의 컴포넌트로 페이지를 구성할지는 axios요청 이후 수정 예정입니다.
