+++ client/src/App.tsx
@@ -3,6 +3,7 @@ import { useNavigate, Outlet } from 'react-router-dom';
 import Footer from './common/footer/Footer';
 import Header from './common/header/Header';
 import Popup from './common/popup/Popup';
+import FeedEditCard from './components/card/feedwritecard/FeedEditCard';
 
 function App() {
   const navigate = useNavigate();
@@ -16,22 +17,7 @@ function App() {
   };
   return (
     <>
-      <Header isloginuser="true" />
-      <div className="w-screen h-screen flex justify-center items-center bg-slate">
-        <Outlet />
-      </div>
-      <Footer />
-      {isOpened && (
-        <Popup
-          title="회원가입을 하셔야 사용합니다"
-          handler={[handleClose, handleNavigate]}
-          btnsize={['md', 'md']}
-          countbtn={2}
-          buttontext={['닫기', '확인']}
-          isgreen={['true', 'false']}
-          popupcontrol={handleClose}
-        />
-      )}
+      <FeedEditCard />
     </>
   );
 }
